Create README.md
# Crowdfunding-ETL
For the ETL mini project, we worked with a partner to practice building an ETL pipeline using Python, Pandas, and Python dictionary
methods or regular expressions to extract and transform the data.  After we transformed the data, we created four CSV files and using 
the CSV file data to create an ERD and table schema.  Finally, we uploaded the CSV file data into a Postgres database.